+++ module/data/item/armor.js
@@ -12,11 +12,21 @@ export class ArmorItemData extends GeneralItemData {
     return {
       ...super.defineSchema('armor'),
       bonus: new fields.NumberField({ initial: 1, integer: true }),
-      equipped: new fields.BooleanField({ initial: false }),
       maxDex: new fields.NumberField({ initial: 12, integer: true }),
       minStrength: new fields.NumberField({ initial: 0, integer: true }),
       notes: new fields.StringField({ initial: '' }),
       fatigue: new fields.NumberField({ initial: 0, integer: true }),
     };
+  }
+
+  static migrateData(source) {
+    if (typeof source.equipped === 'boolean') {
+      source.equipped = { carryType: source.equipped ? 'worn' : 'stowed' };
+    }
+    return super.migrateData(source);
+  }
+
+  get usage() {
+    return 'worn';
   }
 }