+++ module/documents/actor/torgeternityActor.js
@@ -14,11 +14,11 @@ export default class TorgeternityActor extends foundry.documents.Actor {
    * @returns {Item|null}
    */
   get equippedMelee() {
-    return this.itemTypes.meleeweapon.find((a) => a.system.equipped) ?? null;
+    return this.itemTypes.meleeweapon.find((a) => a.isEquipped) ?? null;
   }
 
   get equippedMelees() {
-    return this.itemTypes.meleeweapon.filter((a) => a.system.equipped) ?? null;
+    return this.itemTypes.meleeweapon.filter((a) => a.isEquipped) ?? null;
   }
 
   get race() {
@@ -38,8 +38,8 @@ export default class TorgeternityActor extends foundry.documents.Actor {
     // Here Effects are not yet applied
     if (this.type !== 'vehicle') {
       // initialize the worn armor and shield bonus
-      const wornArmor = this.itemTypes.armor.find((a) => a.system.equipped);
-      const heldShield = this.itemTypes.shield.find((a) => a.system.equipped);
+      const wornArmor = this.itemTypes.armor.find((a) => a.isEquipped);
+      const heldShield = this.itemTypes.shield.find((a) => a.isEquipped);
       const shieldBonus = heldShield?.system?.bonus ?? 0;
 
       this.fatigue = 2 + (wornArmor?.system?.fatigue ?? 0);
@@ -170,9 +170,9 @@ export default class TorgeternityActor extends foundry.documents.Actor {
 
       const meleeWeaponsDefenseSkill = skills.meleeWeapons.value || attributes.dexterity.value;
       this.defenses.meleeWeapons.value = meleeWeaponsDefenseSkill + this.defenses.meleeWeapons.mod;
-      // (Core pg 126) Wielding two melee weapons increases melee weapons defense by 2.
+      // (Core pg 126) Wielding TWO melee weapons increases melee weapons defense by 2.
       if (this.type !== 'vehicle') {
-        const meleeWeaponCount = this.items.filter(item => item.type === 'meleeweapon' && item.system.equipped);
+        const meleeWeaponCount = this.items.filter(item => item.type === 'meleeweapon' && item.isEquipped);
         if (meleeWeaponCount.length > 1) this.defenses.meleeWeapons.value += 2;
       }
 
@@ -681,7 +681,7 @@ export default class TorgeternityActor extends foundry.documents.Actor {
    */
   async setActiveDefense(bonus) {
 
-    const equippedShield = this.items.find(item => item.type === 'shield' && item.system.equipped); // Search for an equipped shield
+    const equippedShield = this.items.find(item => item.type === 'shield' && item.isEquipped); // Search for an equipped shield
     let shieldBonus = (equippedShield && !this.hasStatusEffect('vulnerable') && !this.hasStatusEffect('veryVulnerable')) ? equippedShield.system.bonus : 0
 
     return this.createEmbeddedDocuments('ActiveEffect', [{
@@ -844,7 +844,7 @@ export default class TorgeternityActor extends foundry.documents.Actor {
   get defenseTraits() {
     const result = [];
     for (const item of this.items) {
-      if ((item.type === 'armor' && item.system.equipped) ||
+      if ((item.type === 'armor' && item.isEquipped) ||
         item.type === 'perk' ||
         item.type === 'specialability' ||
         item.type === 'specialabilityRollable') {