+++ module/torgchecks.js
@@ -1045,10 +1045,9 @@ export async function rollAttack(actor, item) {
       switch (attackWith) {
         case 'meleeWeapons':
         case 'unarmedCombat':
-          dnDescriptor = firstTarget.items
-            .filter((it) => it.type === 'meleeweapon' && it.system.equipped).length === 0
-            ? 'targetUnarmedCombat'
-            : 'targetMeleeWeapons';
+          dnDescriptor = firstTarget.items.find(item => item.type === 'meleeweapon' && item.isEquipped)
+            ? 'targetMeleeWeapons'
+            : 'targetUnarmedCombat';
           break;
         case 'fireCombat':
         case 'energyWeapons':
@@ -1217,15 +1216,12 @@ export async function rollUnarmedAttack(actor, skillName) {
     const firstTarget = game.user.targets.find(token => token.actor.type !== 'vehicle')?.actor ||
       game.user.targets.first().actor;
 
-    if (firstTarget.type === 'vehicle') {
+    if (firstTarget.type === 'vehicle')
       dnDescriptor = 'targetVehicleDefense';
-    } else {
-      firstTarget.items
-        .filter((it) => it.type === 'meleeweapon')
-        .filter((it) => it.system.equipped).length !== 0
-        ? (dnDescriptor = 'targetMeleeWeapons')
-        : (dnDescriptor = 'targetUnarmedCombat');
-    }
+    else
+      dnDescriptor = firstTarget.items.find(item => item.type === 'meleeweapon' && item.isEquipped)
+        ? 'targetMeleeWeapons'
+        : 'targetUnarmedCombat';
   }
 
   // Almost the same as rollAttack