+++ Changelog.md
@@ -3,6 +3,7 @@
 ## NEXT
 
 - Add "Test Inspector" context menu option to chat messages.
+- Add "duration=X" option to `@Condition`
 
 ## 13.22.2 - Ammo and other Active Effect fixes
 
@@ -14,7 +15,6 @@
 - AEs on Ammo will no longer generate an error.
 - No error if test has no `attackTraits`.
 - Do not strip `test.X` changes from AEs when copying the AE to a target.
-
 
 **NOTE** : `Transfer on Outcome` really requires two separate entries:
   - Only transfer to target on a specific outcome (single "Transfer on Success": "<blank>", "Any Success" or one of the specific levels)